All attributes using events
All attributes are now pushing events.
In order to reduce the polling and improve the performance, a new
command has been introduced: read_attrs.

This command is the one in charge of reading all the setting attributes
and push the events in case it's needed.

The templates for code generation had been improved in orther to cover
the special diagnostics attributes.